+++ src/main/resources/static/src/app/components/admin/cuadrilla/tabla-cuadrilla/tabla-cuadrilla.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, ViewChild} from '@angular/core';
+import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
 import {Router} from "@angular/router";
 import {Cuadrilla} from "../../../../shared/models/Cuadrilla";
 import {CuadrillaAdminService} from "../../../../shared/services/admin/tabla_cuadrilla/cuadrilla-admin.service";
@@ -6,13 +6,13 @@ import {Proveedor} from "../../../../shared/models/Proveedor";
 import {MatTableDataSource} from "@angular/material/table";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort} from "@angular/material/sort";
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-tabla-cuadrilla',
   templateUrl: './tabla-cuadrilla.component.html',
   styleUrls: ['./tabla-cuadrilla.component.scss']
 })
-export class TablaCuadrillaComponent implements OnInit {
+export class TablaCuadrillaComponent implements OnInit, AfterViewInit {
 
   cuadrilla: Cuadrilla[] = [];
   displayedColumns: string[] = ['ID', 'movil_Asociado','proveedor','turno_trabajo','zona', 'actions'];
@@ -47,10 +47,28 @@ export class TablaCuadrillaComponent implements OnInit {
   }
 
   Delete(cuadrilla: Cuadrilla) {
-    this.tabla_admin_cuadrillascv.deleteCuadrilla(cuadrilla).subscribe(data => {
-      this.cuadrilla = this.cuadrilla.filter(p => p !== cuadrilla);
-      alert("Cuadrilla Eliminada");
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.tabla_admin_cuadrillascv.deleteCuadrilla(cuadrilla).subscribe(data => {
+          this.cuadrilla = this.cuadrilla.filter(p => p !== cuadrilla);
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+          location.href = "/admin/cuadrilla_admin";
+        })
+      }
     })
+
   }
   Agregar(){
