+++ src/main/resources/static/src/app/components/admin/proveedor/tabla-proveedor/tabla-proveedor.component.ts
@@ -6,6 +6,9 @@ import {MatTableDataSource} from "@angular/material/table";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort} from "@angular/material/sort";
 import {MatDialog} from "@angular/material/dialog";
+import {Cuadrilla} from "../../../../shared/models/Cuadrilla";
+import Swal from "sweetalert2";
+import {ProveedorAdminModalComponent } from "../proveedor-admin-modal/proveedor-admin-modal.component";
 
 export interface UserData {
   id: string;
@@ -58,15 +61,48 @@ export class TablaProveedorComponent implements OnInit {
 
   }
 
+
+
   Delete(proveedor: Proveedor) {
-    this.tabla_admin_provvedorscv.deleteProveedor(proveedor).subscribe(data => {
-      this.proveedor = this.proveedor.filter(p => p !== proveedor);
-      alert("Proveedor Eliminado");
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.tabla_admin_provvedorscv.deleteProveedor(proveedor).subscribe(data => {
+          this.proveedor = this.proveedor.filter(p => p !== proveedor);
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+          location.href = "/admin/provedor_admin";
+        })
+      }
     })
+
   }
 
 
   Agregar() {
-
+    this.openDialog();
   }
+  openDialog(proveedor?: Proveedor): void {
+    const config = {
+      data: {
+        message: proveedor ? 'Editar' : 'Agregar',
+        content: proveedor
+      }
+    };
+    const dialogRef = this.dialog.open(ProveedorAdminModalComponent, config);
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(`Dialog result ${result}`);
+    });
+  }
+
 }
